+++ packages/realm-react/CHANGELOG.md
@@ -1,3 +1,16 @@
+x.x.x Release notes (yyyy-MM-dd)
+=============================================================
+### Enhancements
+* None.
+
+### Fixed
+* Adding event listeners while in a write transaction ([#4375](https://github.com/realm/realm-js/issues/4375))
+### Compatibility
+* None.
+
+### Internal
+* None.
+
 0.2.1 Release notes (2022-03-24)
 =============================================================
 ### Enhancements